refactor(markets): Move market pages into components directory
