+++ js/src/pages/markets/components/add-market-modal/index.test.js
@@ -8,7 +8,7 @@ import userEvent from '@testing-library/user-event';
 /**
  * Internal dependencies
  */
-import AddMarketModal from './';
+import AddMarketModal from '.';
 import useSettings from '~/hooks/useSettings';
 import useStoreCurrency from '~/hooks/useStoreCurrency';
 import { SHIPPING_RATE_METHOD } from '~/constants';
