+++ js/src/pages/markets/components/add-market-modal/multilingual-plugin-prompt/index.test.js
