+++ js/src/pages/markets/components/market-data-views/currency-cell.js
