+++ js/src/pages/markets/components/market-data-views/formatted-amount.js
@@ -3,7 +3,7 @@
  */
 import { glaData } from '~/constants';
 import useAdsCurrency from '~/hooks/useAdsCurrency';
-import useMarketCurrency from '../hooks/useMarketCurrency';
+import useMarketCurrency from '../../hooks/useMarketCurrency';
 
 /**
  * Renders a currency-formatted monetary amount.
