+++ js/src/pages/markets/components/market-data-views/index.js
@@ -10,10 +10,10 @@ import { Icon, edit, trash } from '@wordpress/icons';
  * Internal dependencies
  */
 import useTargetAudienceFinalCountryCodes from '~/hooks/useTargetAudienceFinalCountryCodes';
-import useMarketDataViewsConfig from '../hooks/useMarketDataViewsConfig';
+import useMarketDataViewsConfig from '../../hooks/useMarketDataViewsConfig';
 import EditMarketModal from '../edit-market-modal';
 import DeleteMarketModal from '../delete-market-modal';
-import isPrimaryMarket from '../utils/isPrimaryMarket';
+import isPrimaryMarket from '../../utils/isPrimaryMarket';
 import './index.scss';
 
 const DEFAULT_VIEW = {
