+++ js/src/pages/markets/components/market-data-views/index.scss
