+++ js/src/pages/markets/components/market-data-views/language-cell.js
