+++ js/src/pages/markets/components/market-fields/index.scss
