+++ js/src/pages/markets/components/market-fields/locale-section/index.scss
