+++ js/src/pages/markets/components/market-fields/shipping-section/shipping-info-notice/index.scss
