+++ js/src/pages/markets/components/market-fields/shipping-section/shipping-rate-controls/shipping-rate-notice/index.test.js
