+++ js/src/pages/markets/components/markets-dashboard/index.scss
