+++ js/src/pages/markets/components/markets-dashboard/multilingual-flat-shipping-notice/index.scss
